yt_ai_article/requirements.txt
flask
youtube-transcript-api
fpdf
google-generativeai